refactor(settings): extract types + styles + reusable components
Phase 7 first slice. Pull shared infrastructure out of SettingsApp.tsx
into a settings/ subdirectory. Subsequent stack PRs can extract the 8
section panels (General/LocalLLMs/Company/Keys/Integrations/Intervals/
Flags/DangerZone), each importing only from the new shared modules.

  - web/src/components/apps/SettingsApp.tsx: -354 LOC (1792 -> 1438)
  - settings/types.ts: 37 LOC (new) — SectionId, Section, SectionGroup, SectionProps
  - settings/constants.ts: 46 LOC (new) — SECTION_GROUPS sidebar nav
  - settings/styles.ts: 200 LOC (new) — shared inline-style object
  - settings/components.tsx: 101 LOC (new) — Field, SaveButton, KeyField

SettingsApp.tsx falls below the 1500-LOC fail line (now in 800-1500 warn
zone). Allowlist entry removed via the forward-only ratchet — the
scripts/check-file-size.sh stale-entry guard caught it on the first run.

No behavior change. Typecheck clean. All 565 web tests pass. Build clean.

Next slices:
  - extract section panels one-by-one (or grouped as Workspace /
    Credentials / System / Advanced — TBD by reviewer)
  - promote react-hooks/exhaustive-deps to error after section
    extraction surfaces stale-closure bugs
