+++ web/src/components/apps/settings/styles.ts
@@ -1,0 +1,200 @@
+// SettingsApp shared style object. Extracted from SettingsApp.tsx so
+// each section panel and the small components (Field/SaveButton/
+// KeyField) can import the same style tokens without re-defining.
+//
+// Kept as a plain object (not a CSS module) because several entries are
+// functions of state (navItem(active), keyStatus(set)) that compose into
+// inline styles at the call site.
+export const styles = {
+  shell: {
+    display: "flex",
+    flex: 1,
+    minHeight: 0,
+    alignItems: "flex-start",
+  } as const,
+  nav: {
+    width: 260,
+    flexShrink: 0,
+    padding: "14px 12px",
+    position: "sticky" as const,
+    top: 0,
+    maxHeight: "100vh",
+    overflowY: "auto" as const,
+    display: "flex",
+    flexDirection: "column" as const,
+    gap: 14,
+    background: "var(--bg-card)",
+  } as const,
+  navGroupLabel: {
+    fontSize: 10,
+    fontWeight: 600,
+    textTransform: "uppercase" as const,
+    letterSpacing: "0.05em",
+    color: "var(--text-tertiary)",
+    margin: "0 0 4px 10px",
+  } as const,
+  navItem: (active: boolean) => ({
+    display: "flex",
+    alignItems: "center",
+    gap: 10,
+    padding: "8px 10px",
+    fontSize: 12,
+    borderRadius: 6,
+    color: active ? "var(--text)" : "var(--text-secondary)",
+    cursor: "pointer",
+    border: "none",
+    background: active ? "rgba(0, 0, 0, 0.06)" : "transparent",
+    width: "100%",
+    textAlign: "left" as const,
+    fontFamily: "var(--font-sans)",
+    fontWeight: active ? 600 : 400,
+    transition: "all 0.15s",
+  }),
+  navIcon: {
+    width: 16,
+    height: 16,
+    flexShrink: 0,
+    strokeWidth: 2,
+  } as const,
+  body: {
+    flex: 1,
+    padding: "24px 32px",
+    maxWidth: 680,
+  } as const,
+  sectionTitle: { fontSize: 18, fontWeight: 700, marginBottom: 4 } as const,
+  sectionDesc: {
+    fontSize: 13,
+    color: "var(--text-secondary)",
+    marginBottom: 20,
+    lineHeight: 1.5,
+  } as const,
+  banner: {
+    display: "flex",
+    gap: 10,
+    alignItems: "flex-start",
+    padding: "10px 14px",
+    marginBottom: 16,
+    background: "var(--yellow-bg)",
+    borderRadius: "var(--radius-md)",
+    fontSize: 12,
+    lineHeight: 1.5,
+    color: "var(--text)",
+  } as const,
+  row: {
+    display: "flex",
+    alignItems: "flex-start",
+    gap: 12,
+    marginBottom: 14,
+  } as const,
+  rowLabel: { width: 160, flexShrink: 0, paddingTop: 8 } as const,
+  rowLabelName: {
+    fontSize: 13,
+    fontWeight: 500,
+    color: "var(--text)",
+  } as const,
+  rowLabelHint: {
+    fontSize: 11,
+    color: "var(--text-tertiary)",
+    marginTop: 2,
+  } as const,
+  rowField: { flex: 1, minWidth: 0 } as const,
+  input: {
+    background: "var(--bg-card)",
+    border: "1px solid var(--border)",
+    color: "var(--text)",
+    borderRadius: "var(--radius-sm)",
+    height: 36,
+    fontSize: 13,
+    padding: "0 10px",
+    outline: "none",
+    width: "100%",
+    fontFamily: "var(--font-sans)",
+  } as const,
+  textarea: {
+    background: "var(--bg-card)",
+    border: "1px solid var(--border)",
+    color: "var(--text)",
+    borderRadius: "var(--radius-sm)",
+    minHeight: 60,
+    fontSize: 13,
+    padding: "8px 10px",
+    outline: "none",
+    width: "100%",
+    fontFamily: "var(--font-sans)",
+    lineHeight: 1.5,
+    resize: "vertical" as const,
+  },
+  keyStatus: (set: boolean) => ({
+    display: "inline-flex",
+    alignItems: "center",
+    fontSize: 11,
+    fontWeight: 500,
+    padding: "2px 8px",
+    borderRadius: "var(--radius-full)",
+    whiteSpace: "nowrap" as const,
+    background: set ? "var(--green-bg)" : "var(--bg-warm)",
+    color: set ? "var(--green)" : "var(--text-tertiary)",
+  }),
+  saveRow: {
+    display: "flex",
+    gap: 8,
+    marginTop: 20,
+    paddingTop: 16,
+    borderTop: "1px solid var(--border-light)",
+  } as const,
+  filePath: {
+    fontFamily: "var(--font-mono)",
+    fontSize: 11,
+    color: "var(--text-tertiary)",
+    padding: "6px 10px",
+    background: "var(--bg-warm)",
+    borderRadius: "var(--radius-sm)",
+    border: "1px solid var(--border-light)",
+    userSelect: "all" as const,
+    wordBreak: "break-all" as const,
+  },
+  table: {
+    width: "100%",
+    borderCollapse: "collapse" as const,
+    fontSize: 12,
+  } as const,
+  th: {
+    textAlign: "left" as const,
+    fontWeight: 600,
+    fontSize: 11,
+    textTransform: "uppercase" as const,
+    letterSpacing: "0.06em",
+    color: "var(--text-tertiary)",
+    padding: "6px 8px",
+    borderBottom: "1px solid var(--border)",
+  } as const,
+  td: {
+    padding: "6px 8px",
+    borderBottom: "1px solid var(--border-light)",
+    verticalAlign: "top" as const,
+  } as const,
+  tdFlag: {
+    padding: "6px 8px",
+    borderBottom: "1px solid var(--border-light)",
+    verticalAlign: "top" as const,
+    fontFamily: "var(--font-mono)",
+    color: "var(--accent)",
+    whiteSpace: "nowrap" as const,
+  } as const,
+  tdDesc: {
+    padding: "6px 8px",
+    borderBottom: "1px solid var(--border-light)",
+    verticalAlign: "top" as const,
+    color: "var(--text-secondary)",
+  } as const,
+  groupTitle: {
+    fontSize: 11,
+    fontWeight: 600,
+    textTransform: "uppercase" as const,
+    letterSpacing: "0.06em",
+    color: "var(--text-tertiary)",
+    marginBottom: 10,
+    paddingBottom: 6,
+    borderBottom: "1px solid var(--border-light)",
+  } as const,
+};
